gralloc: Fix  implicit conversion changes signedness: 'int' to 'size_t' (aka 'unsigned int')

+++ gralloc/gralloc_priv.h
@@ -74,7 +74,7 @@
 
 #define ROUND_UP_PAGESIZE(x) roundUpToPageSize(x)
 inline size_t roundUpToPageSize(size_t x) {
-  return (x + (getpagesize() - 1)) & ~(getpagesize() - 1);
+  return (x + ((unsigned int)getpagesize() - 1)) & ~((unsigned int)getpagesize() - 1);
 }
 
 /* Legacy gralloc1 definitions */
